Move components to own file; add resource
