+++ src/fastmcp/prompts/prompt.py
@@ -15,11 +15,11 @@ from pydantic import Field, TypeAdapter, validate_call
 
 from fastmcp.exceptions import PromptError
 from fastmcp.server.dependencies import get_context
+from fastmcp.utilities.components import FastMCPComponent
 from fastmcp.utilities.json_schema import compress_schema
 from fastmcp.utilities.logging import get_logger
 from fastmcp.utilities.types import (
     FastMCPBaseModel,
-    FastMCPComponent,
     find_kwarg_by_type,
     get_cached_typeadapter,
 )
